feat: remove result pattern and add permmission

+++ src/Hotelos.Application/Floors/FloorsService.cs
@@ -1,40 +1,42 @@
-﻿using Hotelos.Application.Bases;
+﻿using FluentValidation;
+using Hotelos.Application.Bases;
 using Hotelos.Application.Contracts.Floors;
 using Hotelos.Application.Contracts.Floors.Dtos;
-using Hotelos.Application.Contracts.Results;
 using Hotelos.Application.Floors.Mappers;
 using Hotelos.Application.Floors.Validators;
 using Hotelos.Domain.Rooms.Entities.Floors;
+using Hotelos.Permissions;
 using Microsoft.AspNetCore.Authorization;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Volo.Abp.Caching;
+using Volo.Abp.Domain.Entities;
 using Volo.Abp.Domain.Repositories;
 
 namespace Hotelos.Application.Floors
 {
-    [Authorize]
     public class FloorsService(IRepository<Floor> floorRepository,
                                IDistributedCache<List<GetFloorDto>> floorDistributedCache) : BaseServices, IFloorsService
     {
         private readonly IRepository<Floor> _floorRepository = floorRepository;
         private readonly IDistributedCache<List<GetFloorDto>> _floorDistributedCache = floorDistributedCache;
 
-        public async Task<Result<GetFloorDto>> Create(CreateFloorDto createFloorDtos)
+        [Authorize(HotelosPermissions.CreateFloor)]
+        public async Task<GetFloorDto> Create(CreateFloorDto createFloorDtos)
         {
-            var hotelIdClaim = CurrentUser.FindClaim("hotelId");
-            var hotelId = int.Parse(hotelIdClaim.Value);
-
             var validateResult = new CreateFloorValidator().Validate(createFloorDtos);
             if (!validateResult.IsValid)
             {
-                var result = ValidationErorrResult<GetFloorDto>(validateResult);
-                return result;
+                var message = ValidationErorrResult(validateResult);
+                throw new ValidationException(message);
             }
 
+            var hotelIdClaim = CurrentUser.FindClaim("hotelId");
+            var hotelId = int.Parse(hotelIdClaim.Value);
             Guid userId = (Guid)CurrentUser.Id;
+
             Floor floor = Floor.Create(createFloorDtos.Name,
                                        hotelId,
                                        userId);
@@ -43,21 +45,30 @@ namespace Hotelos.Application.Floors
 
             var mapper = new GetFloorDtoMapper();
             var floorMapping = mapper.ToDto(floor);
-            return Result.Created(floorMapping);
+            return floorMapping;
         }
 
-        public async Task<Result<GetFloorDto>> Update(UpdateFloorDto updateFloorDto)
+        [Authorize(HotelosPermissions.UpdateFloor)]
+        public async Task<GetFloorDto> Update(UpdateFloorDto updateFloorDto)
         {
             var validateResult = new UpdateFloorDtoValidator().Validate(updateFloorDto);
             if (!validateResult.IsValid)
             {
-                var result = ValidationErorrResult<GetFloorDto>(validateResult);
-                return result;
+                var message = ValidationErorrResult(validateResult);
+                throw new ValidationException(message);
             }
 
-            var floor = await _floorRepository.FirstOrDefaultAsync(x => x.Id == updateFloorDto.Id);
-
+            var hotelIdClaim = CurrentUser.FindClaim("hotelId");
+            var hotelId = int.Parse(hotelIdClaim.Value);
             Guid userId = (Guid)CurrentUser.Id;
+
+            var floor = await _floorRepository.FirstOrDefaultAsync(x => x.Id == updateFloorDto.Id && x.HotelId == hotelId);
+
+            if (floor is null)
+            {
+                throw new EntityNotFoundException();
+            }
+
             floor.Update(updateFloorDto.Name, userId);
 
             await _floorRepository.UpdateAsync(floor, true);
@@ -65,21 +76,33 @@ namespace Hotelos.Application.Floors
             var mapper = new GetFloorDtoMapper();
             var floorMapping = mapper.ToDto(floor);
 
-            return Result.Success(floorMapping);
+            return floorMapping;
         }
 
-        public async Task<Result<string>> Delete(int id)
+        [Authorize(HotelosPermissions.DeleteFloor)]
+        public async Task<string> Delete(int id)
         {
-            var floor = await _floorRepository.FirstOrDefaultAsync(x => x.Id == id);
+            var hotelIdClaim = CurrentUser.FindClaim("hotelId");
+            var hotelId = int.Parse(hotelIdClaim.Value);
+
+            var floor = await _floorRepository.FirstOrDefaultAsync(x => x.Id == id && x.HotelId == hotelId);
+            if (floor is null)
+            {
+                throw new EntityNotFoundException();
+            }
+
             await _floorRepository.DeleteAsync(floor, true);
-            return Result.Deleted<string>();
+            return "Success";
         }
 
-        public async Task<Result<List<GetFloorDto>>> GetAll()
+        [Authorize(HotelosPermissions.GetAllFloors)]
+        public async Task<List<GetFloorDto>> GetAll()
         {
-            var cachingResult = await _floorDistributedCache.GetOrAddAsync($"GetAllFloorsOfUser_{CurrentUser.Id}",
+            var hotelIdClaim = CurrentUser.FindClaim("hotelId");
+            var hotelId = int.Parse(hotelIdClaim.Value);
+            var cachingResult = await _floorDistributedCache.GetOrAddAsync($"GetAllFloorsOfUser_{hotelId}",
                                                             async () => await GetAllFromDb());
-            return Result.Success(cachingResult);
+            return cachingResult;
         }
 
         private async Task<List<GetFloorDto>> GetAllFromDb()
